Fixed a double free error while writing binary packages
This error appeared in CentOS 6 while writing binary packages.
The threads may access the same object at the same time and
once a thread freed the object, a second thread will try
to free the object, raising a double-free error and
the segmentation fault associated with it.

Making these variable privates solved the issue.